OVERFLOW-99 Implemented collision detection with respective tiles, such as water and borders. To change the collision property of any of the tile types, change the boolean value for it in BLOCK_TYPES (located in tile.h).
