+++ Tile.cpp
@@ -1,41 +1,7 @@
 ﻿#include <iostream>
 #include "Tile.h"
-#include "ConsoleColors.h"
+
 using namespace std;
-
-//enum BLOCK_TYPE { GRASS, WATER, TREE, FLOOR, BUSH };
-
-struct Block {
-	string name;
-	string description;
-	Color color;
-	bool collides;
-};
-
-const int NUM_BLOCK_TYPES = 19;
-
-
-Block BLOCK_TYPES[NUM_BLOCK_TYPES] = {
-	{"Grass", "just a silly block of grass", DARK_GREEN, false},
-	{"Water", "it's cold to the touch", LIGHT_BLUE, true},
-	{"Bridge", "", BROWN, false},
-	{"Dirt", "what a nice pattern!", TAN, false},
-	{"Bush", "there's pointy thorns in this bush. you probably shouldn't walk through it...", LIME_GREEN, true},
-	{"Fence", "", WHITE, true},
-	{"Enemy", "", RED, true},
-	{"Campfire", "", ORANGE, true},
-	{"Shopkeeper", "", PINK, true},
-	{"Scrummius", "", PURPLE, true},
-	{"Tower Wall", "", GRAY, true},
-	{"Friendly", "", BABY_BLUE, true},
-	{"Friendly", "", LIGHT_GREEN, true},
-	{"Mystery Item", "", LAVENDER, true},
-	{"Old House", "", LIGHT_YELLOW, true},
-	{"Old House", "", DARK_YELLOW, true},
-	{"House Window", "", DARK_PURPLE, true},
-	{"House Door", "", DARK_RED, true},
-	{"Sign", "", BROWN, true}
-};
 
 //Constructors
 Tile::Tile(int r, int c, int tileID)
@@ -85,6 +51,11 @@ int Tile::GetColumn()
 void Tile::SetColumn(int c)
 {
 	col = c;
+}
+
+// Returns the ID number of the tile, designating its type
+int Tile::GetID() {
+	return ID;
 }
 
 //Item getters and setters
