+++ Tile.h
@@ -1,11 +1,46 @@
 #pragma once
 #include <string>
+#include "ConsoleColors.h"
 #include "Item.h"
 #include "NPC.h"
 #include "Enemy.h"
 using namespace std;
 
 const string TILE_DISPLAY = "  ";
+
+//enum BLOCK_TYPE { GRASS, WATER, TREE, FLOOR, BUSH };
+
+struct Block {
+	string name;
+	string description;
+	Color color;
+	bool collides;
+};
+
+const int NUM_BLOCK_TYPES = 19;
+
+
+const Block BLOCK_TYPES[NUM_BLOCK_TYPES] = {
+	{"Grass", "just a silly block of grass", DARK_GREEN, false},
+	{"Water", "it's cold to the touch", LIGHT_BLUE, true},
+	{"Bridge", "", BROWN, false},
+	{"Dirt", "what a nice pattern!", TAN, false},
+	{"Bush", "there's pointy thorns in this bush. you probably shouldn't walk through it...", LIME_GREEN, true},
+	{"Fence", "", WHITE, true},
+	{"Enemy", "", RED, true},
+	{"Campfire", "", ORANGE, true},
+	{"Shopkeeper", "", PINK, true},
+	{"Scrummius", "", PURPLE, true},
+	{"Tower Wall", "", GRAY, true},
+	{"Friendly", "", BABY_BLUE, true},
+	{"Friendly", "", LIGHT_GREEN, true},
+	{"Mystery Item", "", LAVENDER, true},
+	{"Old House", "", LIGHT_YELLOW, true},
+	{"Old House", "", DARK_YELLOW, true},
+	{"House Window", "", DARK_PURPLE, true},
+	{"House Door", "", DARK_RED, true},
+	{"Sign", "", BROWN, true}
+};
 
 /*
 Tile class will hold information for 1 specific tile in the Map
@@ -45,6 +80,9 @@ public:
 	//Column Setter/Getters
 	int GetColumn();
 	void SetColumn(int c);
+
+	// ID Getter
+	int GetID();
   
 	//Item Setter/Getters, passed by reference
 	Item* GetItem();
