+++ startingAreaMap.txt
@@ -43,7 +43,7 @@
 3	3	3	0	0	1	0	0	3	3	0	0	0	0	1	1
 0	0	0	0	1	1	1	0	0	3	3	0	0	0	1	4
 4	0	0	1	1	1	1	1	0	0	3	0	1	1	1	4
-4	0	1	1	1	1	1	1	1	1	1	1	1	0	0	4
+4	0	1	1	1	1	1	1	1	1	2	1	1	0	0	4
 4	0	1	1	1	1	1	1	0	0	3	0	0	0	0	4
 4	0	1	1	1	1	0	0	0	3	3	0	0	0	0	4
 4	0	0	1	1	1	1	0	0	3	0	0	0	8	0	4
